Add styling for AddBookForm
