+++ frontend/src/App.css
@@ -1,3 +1,5 @@
+@import url('https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Cedarville+Cursive&family=Playwrite+DE+LA:wght@100..400&display=swap');
+
 #root {
   width: 100%;
 }
@@ -17,4 +19,16 @@ main {
 
 .align-left {
   align-self: flex-start;
+}
+
+.block {
+  display: block;
+}
+
+h2 {
+  color: #C44900;
+  font-size: 30px;
+  padding: 0;
+  font-family: "Bebas Neue", sans-serif;
+  font-weight: bold;
 }
