+++ frontend/src/components/header/Header.css
@@ -1,5 +1,3 @@
-@import url('https://fonts.googleapis.com/css2?family=Bebas+Neue&family=Cedarville+Cursive&family=Playwrite+DE+LA:wght@100..400&display=swap');
-
 header {
 
     background-color: #183A37;
