+++ frontend/src/pages/BookGalleryPage/bookGalleryPage/BookGalleryPage.css
@@ -1,0 +1,5 @@
+#galleryPage {
+    display: flex;
+    flex-direction: column;
+    gap: 10px;
+}
