+++ frontend/src/pages/BookGalleryPage/bookGalleryPage/BookGalleryPage.tsx
@@ -8,9 +8,9 @@ type BookGalleryPageProps = {
 }
 export default function BookGalleryPage({data}: BookGalleryPageProps) {
     return (
-        <>
-            <Link to={"/books/add"}>Add a Book</Link>
+        <div id={"galleryPage"}>
+            <Link className={"book-button align-right"} to={"/books/add"}>Add a Book</Link>
             <BookGallery data={data} />
-        </>
+        </div>
     );
 }
