+++ frontend/src/pages/BookGalleryPage/components/addBookForm/AddBookForm.css
@@ -1,0 +1,17 @@
+#addForm {
+    display: flex;
+    flex-direction: column;
+    gap: 12px;
+}
+
+#addForm input, select {
+    background-color: #183A37;
+    padding: 8px 5px;
+    border: 1px solid #5b84ea;
+    border-radius: 5px;
+}
+
+#addForm label {
+    margin-bottom: -10px;
+}
+
